backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
python-multipart
pydantic
email-validator
